requirements.txt
deepgram-sdk==v3.10.1
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
